refactor(llama): move SnapshotError
